pyproject.toml
[project]
name = "web-automate-hotel-reservation-ai-agent-mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddg-mcp>=0.1.1",
    "duckduckgo-search>=8.0.4",
    "fastapi>=0.115.13",
    "langchain-anthropic>=0.3.15",
    "langchain-openai>=0.3.23",
    "mcp-use>=1.3.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]
